Upgrade to 0.3.x of es6-module-transpiler.
Note that 0.3.2 seems to have a bug with globals, in which it cannot build
without at least one import named on the command line.  A nonsense import, which
is not actually used in the source, does not affect the output.

It is perfectly fine to remove `imports: {"__nothing" => "__nothing"}` for a
version of the es6-module-transpiler that does not suffer from this problem.